+++ SwissKnife/gcloud/GCloudStorage.py
@@ -1,12 +1,9 @@
 import os
 import logging
 import google.cloud.storage as gcloud
-<<<<<<< HEAD
 from google.cloud.storage.blob import Blob
-=======
 from SwissKnife.info.BucketPath import split_bucket
 
->>>>>>> master
 
 from SwissKnife.info import BUCKET_NAME, BUCKET_PATH_PREFIX
 
@@ -39,7 +36,7 @@ class GCloudStorage:
             self.bucket_name = BUCKET_NAME
             self.bucket_path_prefix = BUCKET_PATH_PREFIX
         else:
-            raise RuntimeError("bucket_name param or BUCKET_PATH environment variable not available.")
+            raise RuntimeError("bucket param or BUCKET_PATH environment variable not available.")
 
         self.storage_client = gcloud.Client()
         self.bucket = self.storage_client.get_bucket(self.bucket_name)
@@ -278,9 +275,9 @@ class GCloudStorage:
         dst_blob = self.bucket.copy_blob(src_blob, destination_bucket, dst_path)
 
         # Log complete paths
-        complete_src_path = self.get_storage_complete_file_path(file_name=source_blob.name, with_bucket=True, with_prefix=True, with_gs=True) 
+        complete_src_path = self.get_storage_complete_file_path(file_name=src_blob.name, with_bucket=True, with_prefix=True, with_gs=True) 
         complete_dst_path = dst_storage.get_storage_complete_file_path(file_name=dst_blob.name, with_bucket=True, with_prefix=True, with_gs=True) 
-        logger.info(f"Blob {complete_src_path} copied to blob {complete_dst_path}")
+        self.logger.info(f"Blob {complete_src_path} copied to blob {complete_dst_path}")
 
         return dst_blob